subspace_farmer::cluster
Module cache

Farming cluster cache

Cache is responsible for caching pieces within allocated space to accelerate plotting and serve pieces in response to DSN requests.

This module exposes some data structures for NATS communication, custom piece cache implementation designed to work with cluster cache and a service function to drive the backend part of the cache.

Structs

ClusterCacheIdentifyBroadcast

Broadcast with identification details by caches

ClusterPieceCache

Cluster cache implementation

Functions

cache_service

Create cache service for specified caches that will be processing incoming requests and send periodic identify notifications

Type Aliases

ClusterCacheIndex

Type alias for cache index used by cluster.