subspace_farmer::single_disk_farm::farming

Module rayon_files

Files abstraction that allows reading concurrently using thread pool

Structs

RayonFiles

Wrapper data structure for multiple files to be used with [rayon] thread pool, where the same file is opened multiple times, once for each thread for faster concurrent reads