subspace_farmer::single_disk_farm::identity

Struct Identity

pub struct Identity { /* private fields */ }

Identity struct is an abstraction of public & secret key related operations.

It is basically a wrapper of the keypair (which holds public & secret keys) and a context that will be used for signing.

Implementations

impl Identity

pub fn file_size() -> usize

Size of the identity file on disk

pub fn open_or_create<B: AsRef<Path>>( base_directory: B, ) -> Result<Self, IdentityError>

Opens the existing identity, or creates a new one.

pub fn open<B: AsRef<Path>>( base_directory: B, ) -> Result<Option<Self>, IdentityError>

Opens the existing identity, returns Ok(None) if it doesn’t exist.

pub fn create<B: AsRef<Path>>(base_directory: B) -> Result<Self, IdentityError>

Creates new identity, overrides identity that might already exist.

pub fn public_key(&self) -> &PublicKey

Returns the public key of the identity.

pub fn secret_key(&self) -> &SecretKey

Returns the secret key of the identity.

pub fn entropy(&self) -> &[u8]

Returns entropy used to generate keypair.

pub fn sign_reward_hash(&self, header_hash: &[u8]) -> Signature

Sign reward hash.