Struct subspace_farmer::single_disk_farm::SingleDiskFarmOptions

pub struct SingleDiskFarmOptions<RC, PF> {

    pub plot_directory: PathBuf,
    pub metadata_directory: PathBuf,
    pub allocated_plotting_space: u64,
    pub farmer_protocol_info: FarmerProtocolInfo,
    pub disk_concurrency: NonZeroU16,
    pub archiving_client: RC,
    pub farming_client: RC,
    pub plot_factory: PF,
    pub reward_address: PublicKey,
    pub bootstrap_nodes: Vec<Multiaddr>,
    pub listen_on: Vec<Multiaddr>,
    pub enable_dsn_archiving: bool,
    pub enable_dsn_sync: bool,
    pub enable_farming: bool,

}

Options for SingleDiskFarm creation

Fields
plot_directory: PathBuf

Path to directory where plots are stored, typically HDD.

metadata_directory: PathBuf

Path to directory for storing metadata, typically SSD.

allocated_plotting_space: u64

How much space in bytes can farm use for plots (metadata space is not included)

farmer_protocol_info: FarmerProtocolInfo

Information about protocol necessary for farmer

disk_concurrency: NonZeroU16

Number of major concurrent operations to allow for disk

archiving_client: RC

Client used for archiving subscriptions

farming_client: RC

Independent client used for farming, such that it is not blocked by archiving

plot_factory: PF

Factory that’ll create/open plot using given options

reward_address: PublicKeybootstrap_nodes: Vec<Multiaddr>

Nodes to connect to on creation, must end with /p2p/QmFoo at the end.

listen_on: Vec<Multiaddr>

List of [Multiaddr] on which to listen for incoming connections.

enable_dsn_archiving: bool

Enable DSN subscription for archiving segments.

enable_dsn_sync: boolenable_farming: bool