Function subspace_farmer::single_disk_plot::farming::audit_sector

pub fn audit_sector<S>(
    public_key: &PublicKey,
    sector_index: u64,
    farmer_protocol_info: &FarmerProtocolInfo,
    global_challenge: &Blake2b256Hash,
    solution_range: SolutionRange,
    sector: S
) -> Result<Option<EligibleSector>, FarmingError> where
    S: Read + Seek, 

Audit a single sector

Note: auditing expects cursor to be set to the beginning of the sector and will move the cursor during its operation. Make sure to return it back to the beginning of the sector if necessary.