Function subspace_farmer::single_disk_plot::plotting::plot_sector

pub async fn plot_sector<PR, S, SM>(
    public_key: &PublicKey,
    sector_index: u64,
    piece_receiver: &PR,
    cancelled: &AtomicBool,
    farmer_protocol_info: &FarmerProtocolInfo,
    sector_output: S,
    sector_metadata_output: SM
) -> Result<PlottedSector, PlotSectorError> where
    PR: PieceReceiver,
    S: Write,
    SM: Write, 

Plot a single sector, where sector and sector_metadata must be positioned correctly (seek to desired offset before calling this function if necessary)

NOTE: Even though this function is async, it has blocking code inside and must be running in a separate thread in order to prevent blocking an executor.