Struct subspace_farmer::single_disk_plot::SingleDiskPlot

pub struct SingleDiskPlot { /* private fields */ }

Single disk plot abstraction is a container for everything necessary to plot/farm with a single disk plot.

Plot starts operating during creation and doesn’t stop until dropped (or error happens).

Implementations

impl SingleDiskPlot

pub fn new<RC>(
options: SingleDiskPlotOptions<RC>
) -> Result<Self, SingleDiskPlotError>where
RC: RpcClient,

Create new single disk plot instance

pub fn collect_summary(directory: PathBuf) -> SingleDiskPlotSummary

Collect summary of single disk plot for presentational purposes

pub fn id(&self) -> &SingleDiskPlotId

ID of this farm

pub fn plotted_sectors_count(&self) -> u64

Number of sectors successfully plotted so far

pub fn plotted_sectors(
&self
) -> impl Iterator<Item = Result<PlottedSector, Error>> + '_

Read information about sectors plotted thus far

pub fn piece_reader(&self) -> PieceReader

Get piece reader to read plot pieces later

pub fn on_sector_plotted(
&self,
callback: Arc<dyn Fn(&PlottedSector) + Send + Sync + 'static>
) -> HandlerId

Subscribe to sector plotting notification

pub fn on_solution(
&self,
callback: Arc<dyn Fn(&SolutionResponse) + Send + Sync + 'static>
) -> HandlerId

Subscribe to new solution notification

pub async fn run(self) -> Result<()>

Run and wait for background threads to exit or return an error

pub fn wipe(directory: &Path) -> Result<()>

Wipe everything that belongs to this single disk plot