Module subspace_farmer::single_plot_farm

Modules

dsn_archiving

Structs

PlotFactoryOptions

SinglePlotFarm

Single plot farm abstraction is a container for everything necessary to plot/farm with a single disk plot.

SinglePlotPieceGetter

SinglePlotPlotter

Enums

PiecesVerificationError

Pieces verification errors.

SinglePlotFarmId

An identifier for single plot farm, can be used for in logs, thread names, etc.

SinglePlotFarmInfo

Important information about the contents of the SinglePlotFarm

SinglePlotFarmSummary

Summary of single plot farm for presentational purposes

SinglePlotPlotterError

Errors that happen during plotting of pieces

Type Definitions

PeerSyncStatusProvider

An alias defining a peer status providing callback.

Trait Aliases

PlotFactory