Module subspace_farmer::ws_rpc_server

Structs

HexPiece

Same as Piece, but serializes/deserialized to/from hex string

HexSha256Hash

Similar to Sha256Hash, but serializes/deserialized to/from hex string

Object

Object stored inside in the history of the blockchain

RpcServerImpl

Farmer RPC server implementation.

Traits

PieceGetter

Something that can be used to get decoded pieces by index

RpcClient

Client implementation for the Rpc RPC API.

RpcServer

Server trait implementation for the Rpc RPC API.