Crate subspace_farmer

subspace-farmer library implementation overview

This library provides droppable/interruptable instances of two processes that can be run in parallel: plotting and farming.

During plotting we create:

a binary plot file, which contains subspace-encoded pieces one after another

a RocksDB commitments database, where key is a tag (first 8 bytes of hmac(encoding, salt)) and value is an offset of corresponding encoded piece in the plot (we can do this because all pieces have the same size).

In short, for every 4096 bytes we also store a record with 8-bytes tag and 8-bytes index (+some overhead of RocksDB itself).

During farming we receive a global challenge and need to find a solution based on target and solution range. In order to find solution, we derive local challenge and use first 8 bytes (the same as tag size) as our target and do range query in RocksDB. For that we interpret target as 64-bit big-endian unsigned integer and find all of the keys in tags database that are target ± ½ * solution range (while also handing overflow/underflow) when interpreted as 64-bit unsigned integers.

Re-exports

pub use jsonrpsee;

Modules

multi_farming

ws_rpc_server

Structs

Archiving

Abstraction around archiving blocks and updating global object map

Commitments

Commitments is a database for commitments.

Farming

Farming structure is an abstraction of the farming process for a single replica plot farming.

Identity

Identity struct is an abstraction of public & secret key related operations.

NodeRpcClient

WsClient wrapper.

ObjectMappings

ObjectMappings is a mapping from arbitrary object hash to its location in archived history.

PiecesToPlot

Collection of pieces that potentially need to be plotted

Plot

Plot is an abstraction for plotted pieces and some mappings.

Enums

ArchivingError

CommitmentError

FarmingError

ObjectMappingError

PlotError

Traits

PlotFile

Trait for mocking plot behaviour

RpcClient

Abstraction of the Remote Procedure Call Client

Functions

plot_pieces

Generates a function that will plot pieces.

retrieve_piece_from_plots

Retrieves and decodes a single piece from multiple plots

Type Definitions

PieceOffset

Index of piece on disk

RpcClientError

To become error type agnostic