Crate subspace_farmer

subspace-farmer library implementation overview

This library provides droppable/interruptable instances of two processes that can be run in parallel: plotting and farming.

During plotting we create:

a binary plot file, which contains subspace-encoded pieces one after another

a RocksDB commitments database, where key is a tag (first 8 bytes of hmac(encoding, salt)) and value is an offset of corresponding encoded piece in the plot (we can do this because all pieces have the same size).

In short, for every piece we also store a record with 8-bytes tag and 8-bytes index (+some overhead of RocksDB itself).

During farming we receive a global challenge and need to find a solution based on target and solution range. In order to find solution, we derive local challenge and use first 8 bytes (the same as tag size) as our target and do range query in RocksDB. For that we interpret target as 64-bit big-endian unsigned integer and find all of the keys in tags database that are target ± ½ * solution range (while also handing overflow/underflow) when interpreted as 64-bit unsigned integers.

Re-exports

pub use jsonrpsee;

Modules

cluster

farm

farmer_cache

node_client

piece_cache

plotter

reward_signing

single_disk_farm

thread_pool_manager

utils

Structs

Identity

Identity struct is an abstraction of public & secret key related operations.

Constants

KNOWN_PEERS_CACHE_SIZE

Size of the LRU cache for peers.