subspace_farmer::cluster::cache

Function cache_service

pub async fn cache_service<C>(
    nats_client: NatsClient,
    caches: &[C],
    cache_group: &str,
    identification_broadcast_interval: Duration,
    primary_instance: bool,
) -> Result<()>
where
    C: PieceCache,

Create cache service for specified caches that will be processing incoming requests and send periodic identify notifications