subspace_farmer::cluster

Module plotter

Farming cluster plotter

Plotter is responsible for plotting sectors in response to farmer requests.

This module exposes some data structures for NATS communication, custom plotter implementation designed to work with cluster plotter and a service function to drive the backend part of the plotter.

Structs

ClusterPlotter

Cluster plotter

Enums

ClusterPlotterId

An ephemeral identifier for a plotter

Functions

plotter_service

Create plotter service that will be processing incoming requests.

Type Aliases

HandlerFn3

Type alias used for event handlers