subspace_farmer

Module farmer_cache

A container that caches pieces

Farmer cache is a container that orchestrates a bunch of piece and plot caches that together persist pieces in a way that is easy to retrieve comparing to decoding pieces from plots.

Structs

FarmerCache

Farmer cache that aggregates different kinds of caches of multiple disks.

FarmerCacheWorker

Farmer cache worker used to drive the farmer cache backend

FarmerCaches

Collection of FarmerCache instances for load balancing