subspace_farmer

Module node_client

Node client abstraction

During farmer operation it needs to communicate with node, for example to receive slot notifications and send solutions to claim rewards.

Implementation is abstracted away behind a trait to allow various implementation depending on use case. Implementation may connect to node via RPC directly, through some kind of networked middleware or even wired without network directly if node and farmer are both running in the same process.

Modules

caching_proxy_node_client

Node client wrapper around another node client that caches some data for better performance and proxies other requests through

rpc_node_client

Node client implementation that connects to node via RPC (WebSockets)

Traits

NodeClient

Abstraction of the Node Client

NodeClientExt

Node Client extension methods that are not necessary for farmer as a library, but might be useful for an app