Module subspace_farmer::single_disk_farm

Primary Farm implementation that deals with hardware directly

Single disk farm is an abstraction that contains an identity, associated plot with metadata and a small piece cache. It fully manages farming and plotting process, including listening to node notifications, producing solutions and singing rewards.

Modules

farming

Farming-related utilities

identity

Farm identity

piece_cache

Cache implementation specific to single disk farm

piece_reader

Piece reader for single disk farm

plot_cache

Plot cache for single disk farm

unbuffered_io_file_windows

Wrapper data structure for unbuffered I/O on Windows

Structs

SingleDiskFarm

Single disk farm abstraction is a container for everything necessary to plot/farm with a single disk.

SingleDiskFarmInfoLock

Exclusive lock for single disk farm info file, ensuring no concurrent edits by cooperating processes is done

SingleDiskFarmOptions

Options used to open single disk farm

Enums

BackgroundTaskError

Errors that happen in background tasks

PlottingError

Errors that happen during plotting

ScrubTarget

Scrub target

SingleDiskFarmError

Errors happening when trying to create/open single disk farm

SingleDiskFarmInfo

Important information about the contents of the SingleDiskFarm

SingleDiskFarmScrubError

Errors happening during scrubbing

SingleDiskFarmSummary

Summary of single disk farm for presentational purposes