Module subspace_farmer::single_disk_farm

Re-exports

pub use crate::single_disk_farm::farming::FarmingError;

Modules

farming

piece_cache

piece_reader

Structs

SingleDiskFarm

Single disk farm abstraction is a container for everything necessary to plot/farm with a single disk.

SingleDiskFarmInfoLock

Exclusive lock for single disk farm info file, ensuring no concurrent edits by cooperating processes is done

SingleDiskFarmOptions

Options used to open single disk farm

Enums

BackgroundTaskError

Errors that happen in background tasks

PlottingError

Errors that happen during plotting

SectorExpirationDetails

Details about sector expiration

SectorPlottingDetails

Details about sector currently being plotted

SectorUpdate

Various sector updates

SingleDiskFarmError

Errors happening when trying to create/open single disk farm

SingleDiskFarmId

An identifier for single disk farm, can be used for in logs, thread names, etc.

SingleDiskFarmInfo

Important information about the contents of the SingleDiskFarm

SingleDiskFarmScrubError

Errors happening during scrubbing

SingleDiskFarmSummary

Summary of single disk farm for presentational purposes