Module subspace_farmer::single_disk_plot

Modules

piece_reader

Structs

SingleDiskPlot

Single disk plot abstraction is a container for everything necessary to plot/farm with a single disk plot.

SingleDiskPlotOptions

Options used to open single dis plot

SingleDiskSemaphore

Semaphore that limits disk access concurrency in strategic places to the number specified during initialization

Enums

BackgroundTaskError

Errors that happen in background tasks

FarmingError

Errors that happen during farming

PlottingError

Errors that happen during plotting

SingleDiskPlotError

Errors happening when trying to create/open single disk plot

SingleDiskPlotId

An identifier for single disk plot, can be used for in logs, thread names, etc.

SingleDiskPlotInfo

Important information about the contents of the SingleDiskPlot

SingleDiskPlotSummary

Summary of single disk plot for presentational purposes