subspace_farmer

Module thread_pool_manager

Thread pool managing utilities for plotting purposes

Structs

PlottingThreadPoolManager

Plotting thread pool manager.

PlottingThreadPoolPair

A wrapper around thread pool pair for plotting purposes

PlottingThreadPoolsGuard

Wrapper around PlottingThreadPoolPair that on Drop will return thread pool back into corresponding PlottingThreadPoolManager.