Module subspace_farmer::utils

Modules

farmer_piece_getter

piece_validator

readers_and_pieces

ss58

Modified version of SS58 parser extracted from Substrate in order to not pull the whole sp-core into farmer application

Structs

AsyncJoinOnDrop

Joins async join handle on drop

Functions

run_future_in_dedicated_thread

Runs future on a dedicated thread with the specified name, will block on drop until background thread with future is stopped too, ensuring nothing is left in memory

tokio_rayon_spawn_handler

This function is supposed to be used with [rayon::ThreadPoolBuilder::spawn_handler()] to inherit current tokio runtime.