Trait subspace_farmer::ws_rpc_server::RpcServer

pub trait RpcServer: Sized + Send + Sync + 'static {
    fn get_piece(
        &self,
        piece_index: PieceIndex
    ) -> Result<Option<HexPiece>, Error>;
    fn find_object(
        &self,
        object_id: HexBlake2b256Hash
    ) -> Result<Option<Object>, Error>;

    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the Rpc RPC API.

Required Methods

fn get_piece(&self, piece_index: PieceIndex) -> Result<Option<HexPiece>, Error>

Get single piece by its index

fn find_object(
&self,
object_id: HexBlake2b256Hash
) -> Result<Option<Object>, Error>

Find object by its ID

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.